# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/add": {
            "post": {
                "security": [
                    {
                        "session": []
                    }
                ],
                "description": "Add new production with specific User",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Add New Product",
                "parameters": [
                    {
                        "description": "Product Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "always returns status 200 but body contains error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vendora.BaseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.ProductEntity"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/edit/{id}": {
            "patch": {
                "security": [
                    {
                        "session": []
                    }
                ],
                "description": "edit a product with specific ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "edit a product",
                "parameters": [
                    {
                        "description": "product input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "always return status 200 but body contains error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vendora.BaseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.ProductEntity"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/tags": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "get products with same tags",
                "parameters": [
                    {
                        "description": "slice of tags",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TagsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "always return status 200 but body contains error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vendora.BaseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ProductEntity"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.ActiveStatus": {
            "type": "string",
            "enum": [
                "DR",
                "SUS",
                "AC"
            ],
            "x-enum-varnames": [
                "Draft",
                "Suspend",
                "Activate"
            ]
        },
...
        "models.ProductEntity": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Active is change from Admin",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ActiveStatus"
                        }
                    ]
                },
                "category_id": {
                    "description": "CategoryID",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "desc": {
                    "description": "Description",
                    "type": "string"
                },
                "discount_price": {
                    "description": "DiscountPrice",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "description": "Images",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "description": "Price is total price",
                    "type": "number"
                },
                "rating": {
                    "description": "Rating",
                    "type": "integer"
                },
                "short_desc": {
                    "description": "ShortDescription",
                    "type": "string"
                },
                "sku": {
                    "description": "SKU is Stock keeping unit",
                    "type": "string"
                },
                "slug": {
                    "description": "Slug",
                    "type": "string"
                },
                "stock": {
                    "description": "Stock",
                    "type": "integer"
                },
                "tags": {
                    "description": "Tags",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "Title",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "description": "UserID",
                    "type": "integer"
                }
            }
        },
        "models.ProductInput": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "discount_price": {
                    "type": "number"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "number"
                },
                "short_desc": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.TagsInput": {
            "type": "object",
            "properties": {
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "vendora.BaseResult": {
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Errors provides list off error that occurred in processing request",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "result": {
                    "description": "Result single/array of any type (object/number/string/boolean) that returns as response"
                },
                "result_count": {
                    "description": "ResultCount specified number of records that returned in result_count field expected result been array.",
                    "type": "integer"
                }
            }
        }
    }
}